Added GDB Enabled Hooks.
The GDB source code contains function calls to gdb_enabled,
which is weakly defined to always return 1.

In some cases, the application binary cannot change from debug
mode to release mode. In those cases, for each component that
is selected debuggable, define a function in that components
space named gdb_enabled that will overwrite the weak function
to something that will check the status of a switch, for example.

Change-Id: Id1a706fcd6e7fff073480fb1c7b4ba3b71e11a3b